requirements.txt
argparse
NumPy
Pillow
